+++ bridgetown-routes/test/helper.rb
@@ -2,3 +2,6 @@
 
 require_relative "../../bridgetown-core/test/helper"
 require "bridgetown-builder"
+require "rack"
+
+Bridgetown.begin!